pyproject.toml
[project]
name = "b3dmar-auth-core"
version = "0.1.0"
description = "Shared auth primitives: JWT, password hashing, RBAC, token revocation"
requires-python = ">=3.12"
dependencies = [
    "python-jose[cryptography]>=3.3,<4",
    "argon2-cffi>=23.1",
    "pydantic>=2.0,<3",
    "fastapi>=0.100",
    "redis>=5.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "fakeredis>=2.21",
    "ruff>=0.8",
    "mypy>=1.13",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/b3dmar_auth"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
